fix: Optimise imports and change filter function emplacement
